Restructure KSH dashboard — move support files into facility_utilization/
queries.py, db.py, m1_ward_forecast.py moved into facility_utilization/ subfolder.
Dashboard file stays in ksh/ root. Imports updated accordingly.
